+++ cmd/vegabenchmark/setup.go
@@ -40,7 +40,6 @@ import (
 	"github.com/cenkalti/backoff"
 	"github.com/golang/mock/gomock"
 	"github.com/golang/protobuf/jsonpb"
-	"github.com/prometheus/common/log"
 )
 
 func setupVega() (*processor.App, processor.Stats, error) {
@@ -130,6 +129,8 @@ func setupVega() (*processor.App, processor.Stats, error) {
 		timeService,
 	)
 
+	epochService := epochtime.NewService(log, epochtime.NewDefaultConfig(), timeService, broker)
+
 	banking := banking.New(
 		log,
 		banking.NewDefaultConfig(),
@@ -140,6 +141,7 @@ func setupVega() (*processor.App, processor.Stats, error) {
 		notary,
 		broker,
 		topology,
+		epochService,
 	)
 
 	genesisHandler := genesis.New(log, genesis.NewDefaultConfig())
@@ -152,8 +154,7 @@ func setupVega() (*processor.App, processor.Stats, error) {
 	timeService.NotifyOnTick(netp.OnChainTimeUpdate)
 	bstats := stats.NewBlockchain()
 
-	epochService := epochtime.NewService(log, epochtime.NewDefaultConfig(), timeService, broker)
-
+	feesTracker := execution.NewFeesTracker(epochService)
 	stateVarEngine := statevar.New(log, statevar.NewDefaultConfig(), broker, topology, commander, timeService)
 	netp.Watch(netparams.WatchParam{
 		Param:   netparams.ValidatorsVoteRequired,
@@ -164,6 +165,7 @@ func setupVega() (*processor.App, processor.Stats, error) {
 		Watcher: stateVarEngine.OnFloatingPointUpdatesDurationUpdate,
 	})
 
+	marketTracker := execution.NewMarketTracker()
 	exec := execution.NewEngine(
 		log,
 		execution.NewDefaultConfig(),
@@ -172,6 +174,8 @@ func setupVega() (*processor.App, processor.Stats, error) {
 		oraclesM,
 		broker,
 		stateVarEngine,
+		feesTracker,
+		marketTracker,
 	)
 
 	netParams := netparams.New(log, netparams.NewDefaultConfig(), broker)
@@ -186,7 +190,7 @@ func setupVega() (*processor.App, processor.Stats, error) {
 		Watcher: delegationEngine.OnMinAmountChanged,
 	})
 
-	rewardEngine := rewards.New(log, rewards.NewDefaultConfig(), broker, delegationEngine, epochService, collateral, timeService, topology)
+	rewardEngine := rewards.New(log, rewards.NewDefaultConfig(), broker, delegationEngine, epochService, collateral, timeService, topology, feesTracker, marketTracker)
 	limits := mocks.NewMockLimits(ctrl)
 	limits.EXPECT().CanTrade().AnyTimes().Return(true)
 	limits.EXPECT().CanProposeMarket().AnyTimes().Return(true)
@@ -212,7 +216,6 @@ func setupVega() (*processor.App, processor.Stats, error) {
 		witness,
 		evtfwd,
 		exec,
-		commander,
 		genesisHandler,
 		governance,
 		notary,
@@ -244,14 +247,7 @@ func setupVega() (*processor.App, processor.Stats, error) {
 
 	// load markets and assets
 	uponGenesisW := func(ctx context.Context, rawstate []byte) error {
-		return uponGenesis(
-			ctx,
-			rawstate,
-			log,
-			assets,
-			collateral,
-			exec,
-		)
+		return uponGenesis(ctx, rawstate, log, assets, collateral, exec)
 	}
 
 	setupGenesis(
@@ -283,11 +279,7 @@ func uponGenesis(
 	}
 
 	for k, v := range state {
-		err := loadAsset(
-			k, types.AssetDetailsFromProto(v),
-			assetSvc, collateral,
-		)
-		if err != nil {
+		if err := loadAsset(log, k, types.AssetDetailsFromProto(v), assetSvc, collateral); err != nil {
 			return err
 		}
 	}
@@ -320,6 +312,7 @@ func uponGenesis(
 }
 
 func loadAsset(
+	log *logging.Logger,
 	id string,
 	v *types.AssetDetails,
 	assets *assets.Service,
@@ -436,6 +429,10 @@ func registerExecutionCallbacks(
 		netparams.WatchParam{
 			Param:   netparams.MarketLiquidityProvisionShapesMaxSize,
 			Watcher: exec.OnMarketLiquidityProvisionShapesMaxSizeUpdate,
+		},
+		netparams.WatchParam{
+			Param:   netparams.MarketMinLpStakeQuantumMultiple,
+			Watcher: exec.OnMinLpStakeQuantumMultipleUpdate,
 		},
 		netparams.WatchParam{
 			Param:   netparams.MarketLiquidityMaximumLiquidityFeeFactorLevel,