+++ nodewallets/commander.go
@@ -68,7 +68,7 @@ func (c *Commander) CommandSync(ctx context.Context, cmd txn.Command, payload pr
 
 func (c *Commander) command(_ context.Context, cmd txn.Command, payload proto.Message, done func(error), ty api.SubmitTransactionRequest_Type) {
 	if c.bc == nil {
-		panic("commander was instantiating without chain")
+		panic("commander was instantiated without a chain")
 	}
 	go func() {
 		ctx, cfunc := context.WithTimeout(context.Background(), 5*time.Second)
@@ -88,10 +88,11 @@ func (c *Commander) command(_ context.Context, cmd txn.Command, payload proto.Me
 		}
 
 		tx := commands.NewTransaction(c.wallet.PubKey().Hex(), marshalledData, signature)
-		_, err = c.bc.SubmitTransactionV2(ctx, tx, ty)
+		_, err = c.bc.SubmitTransactionV2(
+			ctx, tx, api.SubmitTransactionRequest_TYPE_SYNC)
 		if err != nil {
 			// this can happen as network dependent
-			c.log.Debug("could not send transaction to tendermint",
+			c.log.Error("could not send transaction to tendermint",
 				logging.Error(err),
 				logging.String("tx", payload.String()))
 		}