+++ risk/mocks/state_var_mock.go
@@ -5,41 +5,43 @@
 package mocks
 
 import (
-	statevar "code.vegaprotocol.io/vega/types/statevar"
 	context "context"
-	gomock "github.com/golang/mock/gomock"
 	reflect "reflect"
+
+	statevar "code.vegaprotocol.io/vega/types/statevar"
+	gomock "github.com/golang/mock/gomock"
 )
 
-// MockStateVarEngine is a mock of StateVarEngine interface
+// MockStateVarEngine is a mock of StateVarEngine interface.
 type MockStateVarEngine struct {
 	ctrl     *gomock.Controller
 	recorder *MockStateVarEngineMockRecorder
 }
 
-// MockStateVarEngineMockRecorder is the mock recorder for MockStateVarEngine
+// MockStateVarEngineMockRecorder is the mock recorder for MockStateVarEngine.
 type MockStateVarEngineMockRecorder struct {
 	mock *MockStateVarEngine
 }
 
-// NewMockStateVarEngine creates a new mock instance
+// NewMockStateVarEngine creates a new mock instance.
 func NewMockStateVarEngine(ctrl *gomock.Controller) *MockStateVarEngine {
 	mock := &MockStateVarEngine{ctrl: ctrl}
 	mock.recorder = &MockStateVarEngineMockRecorder{mock}
 	return mock
 }
 
-// EXPECT returns an object that allows the caller to indicate expected use
+// EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockStateVarEngine) EXPECT() *MockStateVarEngineMockRecorder {
 	return m.recorder
 }
 
+// NewEvent mocks base method.
 func (m *MockStateVarEngine) NewEvent(arg0, arg1 string, arg2 statevar.StateVarEventType) {
 	m.ctrl.T.Helper()
 	m.ctrl.Call(m, "NewEvent", arg0, arg1, arg2)
 }
 
-// NewEvent indicates an expected call of NewEvent
+// NewEvent indicates an expected call of NewEvent.
 func (mr *MockStateVarEngineMockRecorder) NewEvent(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "NewEvent", reflect.TypeOf((*MockStateVarEngine)(nil).NewEvent), arg0, arg1, arg2)